Moved AegisGrasp cfgs to new file
